홈화면 api db연동 완료된 api

+++ src/main/java/Phonesonal/PhoneBE/web/dto/Home/HomeResultDTO.java
@@ -22,7 +22,7 @@ import java.time.LocalDate;
             private int caloriePercentage;
             private int exercisePercentage;
             private BigDecimal targetWeight;
-            private int currentWeight;
+            private BigDecimal currentWeight;
             private String caloriestatus;
             private String exercisestatus;
             private String comment;
